Several zooming tweaks and fixes
The factor when zooming out was not being computed correctly before. Also,
zooming didn't account for the fact that the padding between the pages is not
zoomed. Finally, the computation of the zoomed scrollTop now uses heights
of the zoomed pages, instead of computing those heights manually using old page
heights. The reason why this helps is that rounding and small division
differences add up to a few pixels, which is noticeable. And those differences
happen because we're multiplying and dividing by 1.1 and 1.5 in a few places.