Work around scaling issues in renderPdf
1) Directly set context._transformMatrix in renderPdf() as a workaround for issue #139.

2) Removed the check if(outputScale != 1) so that scaling works on non-retina screens.
Zooming used to change the resolution of the rendered image without also changing
the size of the image on the screen (so for example zooming out would just make things
blurrier). Also the original fix for #139 screwed up other elements in the document
on non-retina screens without this additional change.